Vacuum & Chamber Infrastructure; CIM; EM
Programmed the basic infrastructure for the vacuum and most necessary functions that will be used. Programmed the Chamber Interaction Model (CIM) which will automatically handle all mixtures and ammo additions when new materials are added to a chamber. Only requires the functions for the interactions to be written. Also implemented an element class and an element manager (EM) which imports element information from a text file and stores them in a list. To add to the elements, simply edit the text file in the resources.